refactored the state objects to put them in the correct folders
